+++ crates/ruff_linter/src/checkers/ast/analyze/expression.rs
@@ -776,8 +776,8 @@ pub(crate) fn expression(expr: &Expr, checker: &mut Checker) {
             if checker.enabled(Rule::NestedMinMax) {
                 pylint::rules::nested_min_max(checker, expr, func, args, keywords);
             }
-            if checker.enabled(Rule::RepeatedKeywords) {
-                pylint::rules::repeated_keywords(checker, call);
+            if checker.enabled(Rule::RepeatedKeywordArgument) {
+                pylint::rules::repeated_keyword_argument(checker, call);
             }
             if checker.enabled(Rule::PytestPatchWithLambda) {
                 if let Some(diagnostic) = flake8_pytest_style::rules::patch_with_lambda(call) {
