+++ crates/ruff_linter/src/rules/pylint/mod.rs
@@ -149,7 +149,10 @@ mod tests {
     #[test_case(Rule::UnnecessaryLambda, Path::new("unnecessary_lambda.py"))]
     #[test_case(Rule::NonAsciiImportName, Path::new("non_ascii_module_import.py"))]
     #[test_case(Rule::NonAsciiName, Path::new("non_ascii_name.py"))]
-    #[test_case(Rule::RepeatedKeywords, Path::new("repeated_keywords.py"))]
+    #[test_case(
+        Rule::RepeatedKeywordArgument,
+        Path::new("repeated_keyword_argument.py")
+    )]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.noqa_code(), path.to_string_lossy());
         let diagnostics = test_path(
