+++ crates/ruff_linter/src/rules/pylint/rules/mod.rs
@@ -44,7 +44,7 @@ pub(crate) use redefined_argument_from_local::*;
 pub(crate) use redefined_loop_name::*;
 pub(crate) use repeated_equality_comparison::*;
 pub(crate) use repeated_isinstance_calls::*;
-pub(crate) use repeated_keywords::*;
+pub(crate) use repeated_keyword_argument::*;
 pub(crate) use return_in_init::*;
 pub(crate) use self_assigning_variable::*;
 pub(crate) use single_string_slots::*;
@@ -117,7 +117,7 @@ mod redefined_argument_from_local;
 mod redefined_loop_name;
 mod repeated_equality_comparison;
 mod repeated_isinstance_calls;
-mod repeated_keywords;
+mod repeated_keyword_argument;
 mod return_in_init;
 mod self_assigning_variable;
 mod single_string_slots;
